fix: isolate migration schema readiness dependencies
